EPMLSTRVEN-33 (get exchange rates)

See merge request epm-lstr/epm-lstr-ven/bank-project!63

+++ client-interface/src/test/java/com/epam/clientinterface/task/UpdateExchangeRatesTest.java
@@ -90,7 +90,7 @@ public class UpdateExchangeRatesTest {
                 }
 
                 when(exchangeRateRepositoryMock.findExchangeRateByCurrencyFromAndCurrencyTo(currencyA, currencyB))
-                    .thenReturn(Optional.of(new ExchangeRate(1L, currencyA, currencyB, RandomUtils.nextDouble())));
+                    .thenReturn(Optional.of(new ExchangeRate((short) 1L, currencyA, currencyB, RandomUtils.nextDouble())));
             }
         }
